CustomDM: First version of the T.A.R. camp xp system.
Kill XP is modified by:
Type: how often a player kills a type of creature
Area: where the player kills a creature
Rest: a general  bonus that charges while the player is idle.
(Hence T.A.R for the initials)

The idea is to encourage the player to diversify what and where he hunts, as farming in one location or the same type of creature for too long will reduce XP earned considerably.

The rest camp xp recovers while the player is not actively hunting so as to make the time travelling between locations or socializing not "worthless" as that just means more xp when he does get back to killing.