CustomDM: Updated Tufa and Yanshi with better versions(public lighting and non-static wells)
CustomDM: Food prices have gone up!

CustomDM: Water and Milk now return the empty flask on consumption. Recycle!

CustomDM: Cows are now milkable!

CustomDM: Players now start with a few flasks of water and some empty ones as a way to tell the player that it's a good and cheap way to recover stamina for starter levels.